LayoutTests/inspector/runtime/promise-native-getter-expected.txt

CONSOLE MESSAGE: line 25: [object Promise]
CONSOLE MESSAGE: line 8: unhandledrejection - "test"
CONSOLE MESSAGE: Unhandled Promise Rejection: test
CONSOLE MESSAGE: line 14: caught - "test"
CONSOLE MESSAGE: line 20: rejectionhandled - "test"
Tests that the injected script only `.then()`s promises if they are returned by native getters.
